Resize window, teleport mouse, and more.
- Added a menu button to reset the window size to match the configured
resolution.
- Added a menu button to enable/disable the ability to teleport the mouse
cursor to the virtual display when clicking over the window.
- Added a menu button to enable/disable the ability to resize the window.
- Increased the maximum possible virtual display resolution to 2K monitors
(3840x2160).
- Added more standard resolutions.
- Removed the previous way to detect whether the mouse is in the virtual
display based on periodic polling.
- Added events to capture the mouse movement and when the window is clicked.
- Enhanced the MouseLocationState to include mouse-related characteristics as
the mouse location, the event generted by the mouse, as well as the screen
where the mouse is positioned, and flags to know the current mouse status
and position, which can be used to later enhance this application.